FSP_jestkie_programisty/requirements/dev.txt
asgiref==3.8.1
black==24.8.0
click==8.1.7
Django==4.2.9
django-ckeditor-5==0.2.15
django-cleanup==9.0.0
django-debug-toolbar==4.4.6
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytz==2024.2
sorl-thumbnail==12.11.0
sort-requirements==1.3.0
sqlparse==0.5.1
transliterate==1.10.2